requirements.txt
paho-mqtt
discord.py
docker
python-dateutil
pytz
redis
requests
beautifulsoup4
